+++ ruuvitag_sensor/ruuvi.py
@@ -3,7 +3,7 @@ import sys
 import os
 import time
 
-from ruuvitag_sensor.url_decoder import UrlDecoder
+from ruuvitag_sensor.decoder import get_decoder
 
 mac_regex = '[0-9a-f]{2}([:])[0-9a-f]{2}(\\1[0-9a-f]{2}){4}$'
 
@@ -56,31 +56,23 @@ class RuuviTagSensor(object):
     @staticmethod
     def convert_data(raw):
         """
-        Convert hexadcimal data to string and validate that data is from RuuviTag.
-        Encoded data part is after ruu.vi/# or r/
+        Validate that data is from RuuviTag and get correct data part.
 
         Returns:
-            string: Encoded sensor data
+            tuple (int, string): Data Format type and Sensor data
         """
-        try:
-            # TODO: Fix conversion so convered data will show https://ruu.vi/# and htts://r/
-            # Now it has e.g. [Non_ASCII characters]ruu.vi/#AjwYAMFc
-            base16_split = [raw[i:i + 2] for i in range(0, len(raw), 2)]
-            selected_hexs = filter(lambda x: int(x, 16) < 128, base16_split)
-            characters = [chr(int(c, 16)) for c in selected_hexs]
-            data = ''.join(characters)
+        # TODO: Check from raw data correct data format
+        data = RuuviTagSensor._get_data_format_2and4(raw)
 
-            # take only part after ruu.vi/# or r/
-            index = data.find('ruu.vi/#')
-            if index > -1:
-                return data[(index + 8):]
-            else:
-                index = data.find('r/')
-                if index > -1:
-                    return data[(index + 2):]
-                return None
-        except:
-            return None
+        if data is not None:
+            return (2, data)
+
+        data = RuuviTagSensor._get_data_format_3(raw)
+
+        if data is not None:
+            return (3, data)
+
+        return (None, None)
 
     @staticmethod
     def find_ruuvitags():
@@ -153,7 +145,7 @@ class RuuviTagSensor(object):
             dict: Latest state
         """
 
-        data = RuuviTagSensor.get_data(self._mac)
+        (data_format, data) = RuuviTagSensor.get_data(self._mac)
 
         if data == self._data:
             return self._state
@@ -163,7 +155,7 @@ class RuuviTagSensor(object):
         if self._data is None:
             self._state = {}
         else:
-            self._state = UrlDecoder().decode_data(self._data)
+            self._state = get_decoder(data_format).decode_data(self._data)
 
         return self._state
 
@@ -196,9 +188,57 @@ class RuuviTagSensor(object):
             # Check MAC whitelist
             if macs and not ble_data[0] in macs:
                 continue
-            encoded = RuuviTagSensor.convert_data(ble_data[1])
+            (data_format, data) = RuuviTagSensor.convert_data(ble_data[1])
             # Check that encoded data is valid RuuviTag data and it is sensor data
-            if encoded is not None:
-                state = UrlDecoder().decode_data(encoded)
+            if data is not None:
+                state = get_decoder(data_format).decode_data(data)
                 if state is not None:
                     yield (ble_data[0], state)
+
+    @staticmethod
+    def _get_data_format_2and4(raw):
+        """
+        Validate that data is from RuuviTag and is Data Format 2 or 4. Convert hexadcimal data to string.
+        Encoded data part is after ruu.vi/# or r/
+
+        Returns:
+            string: Encoded sensor data
+        """
+        try:
+            # TODO: Fix conversion so convered data will show https://ruu.vi/# and htts://r/
+            # Now it has e.g. [Non_ASCII characters]ruu.vi/#AjwYAMFc
+            base16_split = [raw[i:i + 2] for i in range(0, len(raw), 2)]
+            selected_hexs = filter(lambda x: int(x, 16) < 128, base16_split)
+            characters = [chr(int(c, 16)) for c in selected_hexs]
+            data = ''.join(characters)
+
+            # take only part after ruu.vi/# or r/
+            index = data.find('ruu.vi/#')
+            if index > -1:
+                return data[(index + 8):]
+            else:
+                index = data.find('r/')
+                if index > -1:
+                    return data[(index + 2):]
+                return None
+        except:
+            return None
+
+    @staticmethod
+    def _get_data_format_3(raw):
+        """
+        Validate that data is from RuuviTag and is Data Format 3
+
+        Returns:
+            string: Sensor data
+        """
+        try:
+            if len(raw) != 54:
+                return None
+
+            if raw[16:18] != '03':
+                return None
+
+            return raw[16:]
+        except:
+            return None