+++ tests/test_ruuvitag_sensor.py
@@ -45,7 +45,8 @@ class TestRuuviTagSensor(TestCase):
             ('BB:2C:6A:1E:59:3D', 'some other device'),
             ('CC:2C:6A:1E:59:3D', '1E0201060303AAFE1616AAFE10EE037275752E76692F23416A7759414D4663CD'),
             ('DD:2C:6A:1E:59:3D', '1E0201060303AAFE1616AAFE10EE037275752E76692F23416A7759414D4663CD'),
-            ('EE:2C:6A:1E:59:3D', '1F0201060303AAFE1716AAFE10F9037275752E76692F23416A5558314D417730C3')
+            ('EE:2C:6A:1E:59:3D', '1F0201060303AAFE1716AAFE10F9037275752E76692F23416A5558314D417730C3'),
+            ('FF:2C:6A:1E:59:3D', '1902010415FF990403291A1ECE1E02DEF94202CA0B5300000000BB')
         ]
 
         for data in datas:
@@ -57,7 +58,7 @@ class TestRuuviTagSensor(TestCase):
            get_datas)
     def test_find_tags(self):
         tags = RuuviTagSensor.find_ruuvitags()
-        self.assertEqual(4, len(tags))
+        self.assertEqual(5, len(tags))
 
     @patch('ruuvitag_sensor.ble_communication.BleCommunicationDummy.get_datas',
            get_datas)
@@ -75,7 +76,8 @@ class TestRuuviTagSensor(TestCase):
 
     def test_convert_data_not_valid(self):
         encoded = RuuviTagSensor.convert_data('not_valid')
-        self.assertIsNone(encoded)
+        self.assertIsNone(encoded[0])
+        self.assertIsNone(encoded[1])
 
     @patch('ruuvitag_sensor.ble_communication.BleCommunicationDummy.get_datas',
            get_datas)
@@ -84,7 +86,7 @@ class TestRuuviTagSensor(TestCase):
     def test_get_datas(self):
         datas = []
         RuuviTagSensor.get_datas(lambda x: datas.append(x))
-        self.assertEqual(4, len(datas))
+        self.assertEqual(5, len(datas))
 
     @patch('ruuvitag_sensor.ble_communication.BleCommunicationDummy.get_datas',
            get_datas)